pyproject.toml
[tool.poetry]
name = "sustainability-rag"
version = "0.1.0"
description = ""
authors = ["陳尚恩 <angus.chen9001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.2.12"
langchain = "^0.3.11"
langchain-community = "^0.3.11"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.24"
langchain-qdrant = "^0.2.0"
qdrant-client = "^1.12.1"
llama-parse = "^0.5.17"
llama-index = "^0.12.5"
llama-index-core = "^0.12.5"
llama-index-embeddings-openai = "^0.3.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
